refactor: comment에서 wiki를 의존하지 않도록 변경
