+++ maeil-wiki/src/main/java/maeilwiki/comment/CommentRequest.java
@@ -1,11 +1,10 @@
 package maeilwiki.comment;
 
 import maeilwiki.member.Member;
-import maeilwiki.wiki.Wiki;
 
-record CommentRequest(String answer, boolean isAnonymous) {
+public record CommentRequest(String answer, boolean isAnonymous) {
 
-    public Comment toComment(Member member, Wiki wiki) {
-        return new Comment(answer, isAnonymous, member, wiki);
+    public Comment toComment(Member member, Long wikiId) {
+        return new Comment(answer, isAnonymous, member, wikiId);
     }
 }
