+++ maeil-wiki/src/main/java/maeilwiki/comment/CommentService.java
@@ -7,17 +7,14 @@ import java.util.concurrent.ConcurrentHashMap;
 import lombok.RequiredArgsConstructor;
 import maeilwiki.member.Member;
 import maeilwiki.member.MemberRepository;
-import maeilwiki.wiki.Wiki;
-import maeilwiki.wiki.WikiRepository;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import org.springframework.transaction.support.TransactionSynchronizationManager;
 
 @Service
 @RequiredArgsConstructor
-class CommentService {
+public class CommentService {
 
-    private final WikiRepository wikiRepository;
     private final CommentRepository commentRepository;
     private final MemberRepository memberRepository;
     private final CommentLikeRepository commentLikeRepository;
@@ -29,9 +26,7 @@ class CommentService {
         String uuid = UUID.randomUUID().toString();
         Member temporalMember = new Member(uuid, uuid, "GITHUB");
         memberRepository.save(temporalMember);
-        Wiki wiki = wikiRepository.findByIdAndDeletedAtIsNull(wikiId)
-                .orElseThrow(NoSuchElementException::new);
-        Comment comment = request.toComment(temporalMember, wiki);
+        Comment comment = request.toComment(temporalMember, wikiId);
 
         commentRepository.save(comment);
     }
