+++ maeil-wiki/src/main/java/maeilwiki/wiki/WikiApi.java
@@ -1,6 +1,8 @@
 package maeilwiki.wiki;
 
 import lombok.RequiredArgsConstructor;
+import maeilwiki.comment.CommentRequest;
+import maeilwiki.comment.CommentService;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.PathVariable;
@@ -13,6 +15,7 @@ import org.springframework.web.bind.annotation.RestController;
 class WikiApi {
 
     private final WikiService wikiService;
+    private final CommentService commentService;
 
     @PostMapping("/wiki")
     public ResponseEntity<Void> createWiki(@RequestBody WikiRequest request) {
@@ -24,6 +27,32 @@ class WikiApi {
     @DeleteMapping("/wiki/{id}")
     public ResponseEntity<Void> deleteWiki(@PathVariable Long id) {
         wikiService.remove(id);
+
+        return ResponseEntity.noContent().build();
+    }
+
+    @PostMapping("/wiki/{wikiId}/comment")
+    public ResponseEntity<Void> createComment(@RequestBody CommentRequest request, @PathVariable Long wikiId) {
+        wikiService.comment(request, wikiId);
+
+        return ResponseEntity.noContent().build();
+    }
+
+    /**
+     * wikiId는 현재 사용 계획이 없지만, 입력 받는 이유는 다음과 같습니다.
+     * - 나중에 특정 위키가 아카이브 상태로 전환되는 기능을 구현할 때는 필요할 수 있습니다.
+     * - comment가 wiki에 속하는 개념이라는 것을 uri로 표현할 수 있습니다.
+     */
+    @DeleteMapping("/wiki/{wikiId}/comment/{id}")
+    public ResponseEntity<Void> deleteComment(@PathVariable Long wikiId, @PathVariable Long id) {
+        commentService.remove(id);
+
+        return ResponseEntity.noContent().build();
+    }
+
+    @PostMapping("/wiki/{wikiId}/comment/{id}/like")
+    public ResponseEntity<Void> toggleLike(@PathVariable Long wikiId, @PathVariable Long id) {
+        commentService.toggleLike(id);
 
         return ResponseEntity.noContent().build();
     }
