+++ maeil-wiki/src/test/java/maeilwiki/comment/CommentServiceTest.java
@@ -32,16 +32,6 @@ class CommentServiceTest extends IntegrationTestSupport {
     private WikiRepository wikiRepository;
 
     @Test
-    @DisplayName("존재하지 않는 위키에 답변을 작성할 수 없다.")
-    void notfound() {
-        CommentRequest request = new CommentRequest("답변을 작성합니다.", false);
-        Long unknownWikiId = -1L;
-
-        assertThatThrownBy(() -> commentService.comment(request, unknownWikiId))
-                .isInstanceOf(NoSuchElementException.class);
-    }
-
-    @Test
     @DisplayName("존재하지 않는 답변을 삭제할 수 없다.")
     void notFoundComment() {
         Long unknownCommentId = -1L;
@@ -89,7 +79,7 @@ class CommentServiceTest extends IntegrationTestSupport {
         Wiki wiki = new Wiki("question", "backend", false, member);
         wikiRepository.save(wiki);
 
-        Comment comment = new Comment("answer", false, member, wiki);
+        Comment comment = new Comment("answer", false, member, wiki.getId());
         return commentRepository.save(comment);
     }
 }
